// generated/pagopa-proxy/PaymentProblemJson.ts
/**
 * Do not edit this file it is auto-generated by italia-utils / gen-api-models.
 * See https://github.com/teamdigitale/italia-utils
 */
/* tslint:disable */

import { PaymentFault } from "./PaymentFault";
import { withDefault } from "italia-ts-commons/lib/types";
import * as t from "io-ts";
import { WithinRangeInteger } from "italia-ts-commons/lib/numbers";

/**
 * A ProblemJson-like type specific for the GetPayment and ActivatePayment operations
 */

// required attributes
const PaymentProblemJsonR = t.interface({
  detail: PaymentFault
});

// optional attributes
const PaymentProblemJsonO = t.partial({
  type: withDefault(t.string, "about:blank"),

  title: t.string,

  status: WithinRangeInteger(100, 600),

  instance: t.string
});

export const PaymentProblemJson = t.intersection(
  [PaymentProblemJsonR, PaymentProblemJsonO],
  "PaymentProblemJson"
);

export type PaymentProblemJson = t.TypeOf<typeof PaymentProblemJson>;